joe's PLAYER APP dev branch
this branch is for development of the player app's data send and receive methods to the contoller app.  Controller app features a custom 'sendFakeData' method in WebSockets.js for testing event messages to the player app. Development of PlayerApp is focused on connecting to the Controller app, parsing socket messages from the Controller, and loading assets based on those commands.

:wq