Creating New Datatype object of Scenes and Files.
Files are being created as an augmentation of what we were calling assets in the past, these assets, in reality were actually just files, this means that each file object now has its own routes for adding and updating. A file is stored within an asset of a scene wherein a scene can add its own unique handling for any file grouped within. All the old asset routes which were listed previously are now routes for the file.

Scenes are being added as an intermediary data type between Assets and Events, Scenes now contain assets within them. Each event has a variety of scenes which contain an 'order' for playing this means that scenes can dynamically be added and paired with a variety of view types, assets and files depending on the desired output.

A new modal for adding scenes was also created based off the UX document enabling the creation of scenes into events.