Major updates to the events data type, views and routes.
Overhauling the events objects include views/single and views/index, routes and modals for events and assets to play nicely with our new data structure of events, scenes, assets, files. These changes reflect new data parsing and collection as well as a new structure for information to the viewer from the application server on load.

Pages such as index and single have been modified to conform with the updated UX/UI documents, additionally modals were shifted slightly in order to facilitate the proper routing of post data through the site.