Updating the front end JS to support new datatypes, adding a sceneController for the /views/event/single.ejs page
The Scene Controller serves to allow the real time manipulation and editing of scenes in the frontend media manager through jquery. The call is loaded into the footer of the page by passing through the current scene object as parsed via the server routes. This allows the frontend to be manipulated with realtime data without a second ajax call to the server facilitating all the data to be loaded at one time. See footer.ejs for SceneController init functionality.