Updating Header and Footer to have more robust features for serving UI
Routes pages have been updated with locals.current allowing each page to know what is currently being served to the user. This benefits bothheader and footer files as it allows us to dynamically edit the headerbar to determine which page we are currently viewing and show our users. It also facilitates the use of custom JS loading in the footer, which means that only the necessary resources for each page are loaded. Meaning that Modals, and js which are unused to not waste memory in the stack being pulled into the page/downloaded by the user