Modifying Modules on the serverside to have the functionality we need.
Utils - added delete file function with async callback to allow for temp files created by the server to be destroyed so that there is no memory leak.

FolderStructure - Fixing minor bug reported by joe for the folder structure start up - should work 100% by searching for the correct file paths and then if they are not located by creating that file structure

DbConnection - massive overhaul to support scenes and assets and files, in particular exports.formatScenes - allows for a prepopulated object of scenes with assets and files to be compiled and return. Also adding function for removing documents from the DB. Adding appropriate collection to the collections object. Updated the make mongo id to be a global as well as reference a local function. the function now enables live checking of string length and data type to determine if we need to create a new mongo object or not. Lastly making global getDocumentbyID to enable routes to search for documents in addition to the local version of getDocumentbyID