Updating event-single page to support no scenes and no assets in scenes.
An error occured previously when trying to load the event/single route when there wasn't any scenes contained. events.js was updated to allow for the page to load content even if the event does not contain any scenes. Additionally, functionality was placed into footer.ejs so that when a user loads a page, which does not contain any scenes yet, the Add New Scenes Modal automattically appears prompting users to add a modal.

The single.ejs page was updated in order to support this functionality, now when looping through the assets, it checks if a scene has any assets store yet. If there are no stored assets, then the page loads placeholder assets for the user to manage. If the scene already contains asset data, that data is loaded directly into the page as was the functionality previously.

Additonally, all these files have been migrated to support the zone_type as an int not as zone_(int) this means that the conditional logic checks in both the single.ejs as well as the sceneController had to be updated to support the mongodb / form data, which comes back to us (the front end) as a string not an int.