Updating sceneController to handle creation of Assets based on Scene DB Information.
Modifying the core class of createAssets as a means of modifying the asset structure in the events/single.ejs file. This allows a reference point back to the db scene objects as well as logic to determine which scene is currently being focused on. As the scenes are switched the createAssets function now manipulates html to create the appropriate number of scene assets and if the scene has assets logged in our database object it creates the appropriate assets in the front end.