Updating files in order to support new locations as a db object.
In adding new file from the new_asset modal - data post reads new location and queries or determine if the location should be added to the database. If the location exists it uses the existing location._id otherwise, it creates a new location and then uses the location._id. In order to clean up the file a new local method was added to handle the file/asset uploading and then storing of the information in the database. function handleFile() now sorts the image type and then uploads the file, function addNewFile is then called within handleFile() and serves the response to the user.