Fixing Bug on loading front end assets
Issue occurred in the current scene change on the frontend, when a user clicks a to load a new scene. The asset/zone paring location algorithm was askew and would result in one or more assets not recieving the data. The fix was created by relying more on the asset key in addition to the adding of a second for loop.

A second loop ensures that all the appropriate div elements are made, it then runs through each of the assets in the scene object and populates all the assets, because the indexs are out of sync the second loop facilitates the correct population of the html. Additionally the asset key object allows jquery to locate and manipulate the correct form item.