Fixing HTML bug which prevented button from being selected.
Previously the button was wrapped around a small tag, and pushed to the side of the div. The button element was no longer accesible via the mouse, and therefore could not activate the new_scene modal. By moving the <small> inside of the <button> we were able to combat this, however it looks ugly. It will need some styling.