Adding API Routes and functionality.
Creating the endpoints for the Savant API to control the PlayerApp. Modified Websockets to recieve and send on the socket only when connected, to prevent failing. Added recall into routes/api.js in order to facilitate messages printing into the browser( for the time being) production code should use the modified, responses of success or error as current commented out.