Fixing Delete Event Bug!
Undocumented bug, which did not delete scenes from the database, when an event was removed.

The architecture has been updated to clean up after itself by first getting the event object to be deleted, checking if it has any scenes, deleting all the scenes from the DB associated with the event and then removing the event object from the database.