Fixing bug in api.js route
bug existed where api route would not respond to server request if no ambient or sales content existed. The fix was the count the size of the object if its 0 then move on to the next element or respond to the user. If its greater than 0 process the data into our API structure.