Adding support for scene reodering on event/single view
Jquery handler within sceneController.js listens for clicks on the button.scene-order when the action is triggered, jquery checks the element for classes to determine if the scene should move up or down in order. From there the sceneController creates a temporary form with the appropriate action and method point to /scenes/reorder/:id.

The scenes router then parses the incoming data, determines the direction passed from jquery reorganizes the array and exists. This functionality only works on one item.