Implimenting Updating and deleting strategies for files
Including a new js file which is responsible for handling the file index page. Flushing out the update route to support updates without a new file and with a new file. Retailored the architecture of handleFile to take a bool option at the end which signals an update vs new file creation.

**Enjoy Kids.**