Adding more robust UI/UX support for options of locations vs files and multi-text vs files
When a user has a file selected and makes a change to show a location, the zone image updates accordingly. Vice versa when a location is set and a file is selected the map is removed, and the selection updated. The same logic handling occurs for the multi-text options, when a user selects multitext they cannot have an image and the UI prevents this and updates the image with holder.js