Updates to render functionality
Basic methods for surfacing and binding render queue by surfacing clips that are flagged for render and then creating an object of scenes with all the associated clips and files. Adding the appropriate collections into the DBConnection to allow for us to have a holder class for our AE template Data objects.