Fixing Scene Add Bug
When a new scene is added we must determine the order of the scene, by default this order is presented as 1,2,3,4 etc. when we check the length of the scene, if the length is 0 we need to set it to 1 otherwise our order ends up 0,2,3,4,5. This has been fixed by adding an inline conditional when we create our post.order.