Adding fix for File deletion issue for AJAX uploaded images
When an image was uploaded via ajax, it did not contain a slug field. When the file/index.ejs was called the slug was undefined, and on delete the file could not be found to be deleted. This issue (#120) is corrected as of this commit. By adding the slug field to the routes/files AJAX handler.