Adding support for rib vertex following in default_scene.jsx
This change allows for the pathfinding to stay the way it is, without significant changes, however it adds greater functionality to the pathfinding, by searching through the vertices of the Custom Mask of the structure to derive points and create a geolocating system.

Also updating the custom composition object to be an AEHelper Object with cleaner functionality it allows for more uses.